+++ apps/web/src/app/layout.tsx
@@ -1,0 +1,27 @@
+import type { Metadata } from 'next';
+import './globals.css';
+import localFont from 'next/font/local';
+
+const pretendard = localFont({
+  src: '../../public/PretendardVariable.woff2',
+  display: 'swap',
+  weight: '45 920',
+  variable: '--font-pretendard',
+});
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="kr" className={`${pretendard.variable}`}>
+      <body className={pretendard.className}>{children}</body>
+    </html>
+  );
+}
